requirements.txt
# MiniNutriScan 项目依赖包列表
# 使用命令安装: pip install -r requirements.txt

# ===========================================
# Web框架和API
# ===========================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ===========================================
# 数据库相关
# ===========================================
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# ===========================================
# Redis缓存
# ===========================================
redis==5.0.1

# ===========================================
# 身份验证和安全
# ===========================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# ===========================================
# HTTP请求和API调用
# ===========================================
requests==2.31.0
httpx==0.25.2

# ===========================================
# 数据处理和验证
# ===========================================
pydantic==2.5.0
pydantic-settings==2.1.0

# ===========================================
# 图像处理
# ===========================================
Pillow==10.1.0

# ===========================================
# 环境变量管理
# ===========================================
python-dotenv==1.0.0

# ===========================================
# 日期时间处理
# ===========================================
python-dateutil==2.8.2

# ===========================================
# 日志和调试
# ===========================================
loguru==0.7.2

# ===========================================
# 开发和测试工具
# ===========================================
pytest==7.4.3
pytest-asyncio==0.21.1

# ===========================================
# 云服务SDK（可选）
# ===========================================
# 阿里云SDK
alibabacloud-ocr-api20210707==1.0.10
alibabacloud-tea-openapi==0.3.7

# 腾讯云SDK
tencentcloud-sdk-python==3.0.1056

# ===========================================
# 其他工具
# ===========================================
aiofiles==23.2.1
jinja2==3.1.2